move prompt into internal package
